Variables pi and e

+++ Starter/Lesson2(Console APP)/Lesson2(Console APP)/Program.cs
@@ -7,7 +7,7 @@ namespace Lesson2_Console_APP_
         static void Main()
         {
             double pi = 3.14159265;
-            double e = 2.7182818284590452;
+            decimal e = 2.7182818284590452m;
             Console.WriteLine(pi);
             Console.WriteLine(e);
             Console.ReadKey();
